+++ scripts/build-standalone.sh
@@ -91,7 +91,7 @@ cp "$HERE/assets/faceid-icon.png" "$RES/faceid-icon.png"
 cp "$HERE/assets/menubar-icon.png" "$RES/menubar-icon.png"
 cp -R "$HERE/packaging/dist/faceid" "$RES/faceid"                 # moteur Python autonome
 mkdir -p "$RES/helpers" "$RES/assets" "$RES/models" "$RES/pam" "$RES/scripts"
-cp "$HERE/helpers/touchid-helper" "$HERE/helpers/auth-modal" "$HERE/helpers/faceid-hud" "$RES/helpers/"
+cp "$HERE/helpers/touchid-helper" "$HERE/helpers/auth-modal" "$HERE/helpers/faceid-hud" "$HERE/helpers/camera-list" "$RES/helpers/"
 cp "$HERE/assets/faceid-icon.png" "$RES/assets/"
 cp "$HERE/assets/FaceID.icns" "$RES/assets/faceid-icon.icns"
 cp "$MODELS/"*.onnx "$RES/models/"